+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Replace npm-shrinkwrap.json with package-lock.json
 * Update packages
 * Support the asynchronous script execution
+* Support arguments passed to a script to be run
 
 ## 1.2.0 (2017-06-03)